+++ Runtime/MemoryManagement/IPool.cs
@@ -2,13 +2,13 @@ namespace RishUI.MemoryManagement
 {
     internal interface IPool
     {
-        uint GetFreeID<T>() where T : class, IManaged, new();
+        ulong GetFreeID<T>() where T : class, IManaged, new();
         
-        T GetManaged<T>(uint id) where T : class, IManaged;
+        T GetManaged<T>(ulong id) where T : class, IManaged;
         
         void CleanGarbage();
 
-        void RegisterReference(uint id, IOwner owner);
-        void UnregisterReference(uint id, IOwner owner);
+        void RegisterReference(ulong id, IOwner owner);
+        void UnregisterReference(ulong id, IOwner owner);
     }
 }
